Move preprocessing functions
